pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "solana-trader"
version = "0.1.0"
description = "Solana arbitrage scanner — Jupiter-backed detector with latency-tracked pipeline (Phase 1)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "python-dotenv>=1.0",
    "psycopg2-binary>=2.9",
    "base58>=2.1",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
# Phase 3+: Solana execution stack.  Not required for the scanner.
execution = [
    "solders>=0.21",
    "solana>=0.34",
    "base58>=2.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
